Refactored formatters to separate files
